move KV and CDI to the new operand pkg
