+++ pkg/controller/common/consts.go
@@ -1,0 +1,6 @@
+package common
+
+const (
+    ReconcileCompleted          = "ReconcileCompleted"
+    ReconcileCompletedMessage   = "Reconcile completed successfully"
+)
