+++ pkg/controller/hyperconverged/hyperconverged_controller.go
@@ -3,10 +3,12 @@ package hyperconverged
 import (
 	"errors"
 	"fmt"
-	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/common"
 	"os"
 	"reflect"
 
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/common"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/operands"
+	
 	"github.com/operator-framework/operator-sdk/pkg/ready"
 	schedulingv1 "k8s.io/api/scheduling/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -30,10 +32,8 @@ import (
 	vmimportv1beta1 "github.com/kubevirt/vm-import-operator/pkg/apis/v2v/v1beta1"
 	conditionsv1 "github.com/openshift/custom-resource-status/conditions/v1"
 	objectreferencesv1 "github.com/openshift/custom-resource-status/objectreferences/v1"
-	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
 
 	corev1 "k8s.io/api/core/v1"
-	rbacv1 "k8s.io/api/rbac/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -77,8 +77,6 @@ const (
 	metricsAggregationOldCrdName    = "kubevirtmetricsaggregations.kubevirt.io"
 	nodeLabellerBundlesOldCrdName   = "kubevirtnodelabellerbundles.kubevirt.io"
 	templateValidatorsOldCrdName    = "kubevirttemplatevalidators.kubevirt.io"
-
-	kubevirtDefaultNetworkInterfaceValue = "masquerade"
 )
 
 // Add creates a new HyperConverged Controller and adds it to the Manager. The Manager will set fields on the Controller
@@ -87,6 +85,16 @@ func Add(mgr manager.Manager, ci hcoutil.ClusterInfo) error {
 	return add(mgr, newReconciler(mgr, ci))
 }
 
+// temp map, until we move all the operands code
+var operandMap = map[string]operands.Operand{}
+func prepareHandlerMap(clt client.Client, scheme *runtime.Scheme) {
+	operandMap["kvc"] = &operands.KvConfigHandler{Client: clt, Scheme: scheme}
+	operandMap["kvpc"] = &operands.KvPriorityClassHandler{Client: clt, Scheme: scheme}
+	operandMap["kv"] = &operands.KubevirtHandler{Client: clt, Scheme: scheme}
+	operandMap["cdi"] = &operands.CdiHandler{Client: clt, Scheme: scheme}
+}
+
+
 // newReconciler returns a new reconcile.Reconciler
 func newReconciler(mgr manager.Manager, ci hcoutil.ClusterInfo) reconcile.Reconciler {
 
@@ -94,6 +102,8 @@ func newReconciler(mgr manager.Manager, ci hcoutil.ClusterInfo) reconcile.Reconc
 	if ownVersion == "" {
 		ownVersion = version.Version
 	}
+
+	prepareHandlerMap(mgr.GetClient(), mgr.GetScheme())
 
 	return &ReconcileHyperConverged{
 		client:      mgr.GetClient(),
@@ -305,7 +315,7 @@ func (r *ReconcileHyperConverged) doReconcile(req *common.HcoRequest) (reconcile
 		req.Logger.Info(fmt.Sprintf("Start upgrating from version %s to version %s", knownHcoVersion, r.ownVersion))
 	}
 
-	req.ComponentUpgradeInProgress = r.upgradeMode
+	req.SetUpgradeMode(r.upgradeMode)
 
 	r.ensureConsoleCLIDownload(req)
 
@@ -453,7 +463,7 @@ func (r *ReconcileHyperConverged) ensureHcoDeleted(req *common.HcoRequest) (reco
 }
 
 func (r *ReconcileHyperConverged) ensureHco(req *common.HcoRequest) error {
-	for _, f := range []func(*common.HcoRequest) *EnsureResult{
+	for _, f := range []func(*common.HcoRequest) *operands.EnsureResult{
 		r.ensureKubeVirtPriorityClass,
 		r.ensureKubeVirtConfig,
 		r.ensureKubeVirt,
@@ -758,296 +768,26 @@ func (r *ReconcileHyperConverged) checkComponentVersion(versionEnvName, actualVe
 	return expectedVersion != "" && expectedVersion == actualVersion
 }
 
-func newKubeVirtConfigForCR(cr *hcov1beta1.HyperConverged, namespace string) *corev1.ConfigMap {
-	labels := map[string]string{
-		hcoutil.AppLabel: cr.Name,
-	}
-	cm := &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kubevirt-config",
-			Labels:    labels,
-			Namespace: namespace,
-		},
-		// only virtconfig.SmbiosConfigKey, virtconfig.MachineTypeKey, virtconfig.SELinuxLauncherTypeKey,
-		// virtconfig.FeatureGatesKey and virtconfig.UseEmulationKey are going to be manipulated
-		// and only on HCO upgrades.
-		// TODO: This is going to change in the next HCO release where the whole configMap is going
-		// to be continuously reconciled
-		Data: map[string]string{
-			virtconfig.FeatureGatesKey:        "DataVolumes,SRIOV,LiveMigration,CPUManager,CPUNodeDiscovery,Sidecar,Snapshot",
-			virtconfig.MigrationsConfigKey:    `{"nodeDrainTaintKey" : "node.kubernetes.io/unschedulable"}`,
-			virtconfig.SELinuxLauncherTypeKey: "virt_launcher.process",
-			virtconfig.NetworkInterfaceKey:    kubevirtDefaultNetworkInterfaceValue,
-		},
-	}
-	val, ok := os.LookupEnv("SMBIOS")
-	if ok && val != "" {
-		cm.Data[virtconfig.SmbiosConfigKey] = val
-	}
-	val, ok = os.LookupEnv("MACHINETYPE")
-	if ok && val != "" {
-		cm.Data[virtconfig.MachineTypeKey] = val
-	}
-	val, ok = os.LookupEnv("KVM_EMULATION")
-	if ok && val != "" {
-		cm.Data[virtconfig.UseEmulationKey] = val
-	}
-	return cm
+func (r *ReconcileHyperConverged) ensureKubeVirtPriorityClass(req *common.HcoRequest) *operands.EnsureResult {
+	return operandMap["kvpc"].Ensure(req)
 }
 
-func (r *ReconcileHyperConverged) ensureKubeVirtConfig(req *common.HcoRequest) *EnsureResult {
-	kubevirtConfig := newKubeVirtConfigForCR(req.Instance, req.Namespace)
-	res := NewEnsureResult(kubevirtConfig)
-	err := controllerutil.SetControllerReference(req.Instance, kubevirtConfig, r.scheme)
-	if err != nil {
-		return res.Error(err)
-	}
-
-	key, err := client.ObjectKeyFromObject(kubevirtConfig)
-	if err != nil {
-		req.Logger.Error(err, "Failed to get object key for kubevirt config")
-	}
-	res.SetName(key.Name)
-
-	found := &corev1.ConfigMap{}
-	err = r.client.Get(req.Ctx, key, found)
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			req.Logger.Info("Creating kubevirt config")
-			err = r.client.Create(req.Ctx, kubevirtConfig)
-			if err == nil {
-				return res.SetCreated()
-			}
-		}
-		return res.Error(err)
-	}
-
-	req.Logger.Info("KubeVirt config already exists", "KubeVirtConfig.Namespace", found.Namespace, "KubeVirtConfig.Name", found.Name)
-	// Add it to the list of RelatedObjects if found
-	objectRef, err := reference.GetReference(r.scheme, found)
-	if err != nil {
-		return res.Error(err)
-	}
-	objectreferencesv1.SetObjectReference(&req.Instance.Status.RelatedObjects, *objectRef)
-
-	if r.upgradeMode {
-		// only virtconfig.SmbiosConfigKey, virtconfig.MachineTypeKey, virtconfig.SELinuxLauncherTypeKey,
-		// virtconfig.FeatureGatesKey and virtconfig.UseEmulationKey are going to be manipulated
-		// and only on HCO upgrades.
-		// TODO: This is going to change in the next HCO release where the whole configMap is going
-		// to be continuously reconciled
-		for _, k := range []string{
-			virtconfig.FeatureGatesKey,
-			virtconfig.SmbiosConfigKey,
-			virtconfig.MachineTypeKey,
-			virtconfig.SELinuxLauncherTypeKey,
-			virtconfig.UseEmulationKey,
-		} {
-			if found.Data[k] != kubevirtConfig.Data[k] {
-				req.Logger.Info(fmt.Sprintf("Updating %s on existing KubeVirt config", k))
-				found.Data[k] = kubevirtConfig.Data[k]
-				err = r.client.Update(req.Ctx, found)
-				if err != nil {
-					req.Logger.Error(err, fmt.Sprintf("Failed updating %s on an existing kubevirt config", k))
-					return res.Error(err)
-				}
-			}
-		}
-	}
-
-	return res.SetUpgradeDone(req.ComponentUpgradeInProgress)
+func (r *ReconcileHyperConverged) ensureKubeVirtConfig(req *common.HcoRequest) *operands.EnsureResult {
+	return operandMap["kvc"].Ensure(req)
 }
 
-func (r *ReconcileHyperConverged) ensureKubeVirtPriorityClass(req *common.HcoRequest) *EnsureResult {
-	req.Logger.Info("Reconciling KubeVirt PriorityClass")
-	pc := req.Instance.NewKubeVirtPriorityClass()
-	res := NewEnsureResult(pc)
-	key, err := client.ObjectKeyFromObject(pc)
-	if err != nil {
-		req.Logger.Error(err, "Failed to get object key for KubeVirt PriorityClass")
-		return res.Error(err)
-	}
-
-	res.SetName(key.Name)
-	found := &schedulingv1.PriorityClass{}
-	err = r.client.Get(req.Ctx, key, found)
-
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			// create the new object
-			err = r.client.Create(req.Ctx, pc, &client.CreateOptions{})
-			if err == nil {
-				return res.SetCreated()
-			}
-		}
-
-		return res.Error(err)
-	}
-
-	// at this point we found the object in the cache and we check if something was changed
-	if pc.Name == found.Name && pc.Value == found.Value && pc.Description == found.Description {
-		req.Logger.Info("KubeVirt PriorityClass already exists", "PriorityClass.Name", pc.Name)
-		objectRef, err := reference.GetReference(r.scheme, found)
-		if err != nil {
-			req.Logger.Error(err, "failed getting object reference for found object")
-			return res.Error(err)
-		}
-		objectreferencesv1.SetObjectReference(&req.Instance.Status.RelatedObjects, *objectRef)
-
-		return res.SetUpgradeDone(req.ComponentUpgradeInProgress)
-	}
-
-	// something was changed but since we can't patch a priority class object, we remove it
-	err = r.client.Delete(req.Ctx, found, &client.DeleteOptions{})
-	if err != nil {
-		return res.Error(err)
-	}
-
-	// create the new object
-	err = r.client.Create(req.Ctx, pc, &client.CreateOptions{})
-	if err != nil {
-		return res.Error(err)
-	}
-	return res.SetUpdated()
+func (r *ReconcileHyperConverged) ensureKubeVirt(req *common.HcoRequest) *operands.EnsureResult {
+	return operandMap["kv"].Ensure(req)
 }
 
-func (r *ReconcileHyperConverged) ensureKubeVirt(req *common.HcoRequest) *EnsureResult {
-	virt := req.Instance.NewKubeVirt()
-	res := NewEnsureResult(virt)
-	if err := controllerutil.SetControllerReference(req.Instance, virt, r.scheme); err != nil {
-		return res.Error(err)
-	}
-
-	key, err := client.ObjectKeyFromObject(virt)
-	if err != nil {
-		req.Logger.Error(err, "Failed to get object key for KubeVirt")
-	}
-
-	res.SetName(key.Name)
-	found := &kubevirtv1.KubeVirt{}
-	err = r.client.Get(req.Ctx, key, found)
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			req.Logger.Info("Creating kubevirt")
-			err = r.client.Create(req.Ctx, virt)
-			if err == nil {
-				return res.SetCreated().SetName(virt.Name)
-			}
-		}
-		return res.Error(err)
-	}
-
-	req.Logger.Info("KubeVirt already exists", "KubeVirt.Namespace", found.Namespace, "KubeVirt.Name", found.Name)
-
-	if !reflect.DeepEqual(found.Spec, virt.Spec) {
-		virt.Spec.DeepCopyInto(&found.Spec)
-		req.Logger.Info("Updating existing KubeVirt's Spec to its default value")
-		err = r.client.Update(req.Ctx, found)
-		if err != nil {
-			return res.Error(err)
-		}
-		return res.SetUpdated()
-	}
-
-	// Add it to the list of RelatedObjects if found
-	objectRef, err := reference.GetReference(r.scheme, found)
-	if err != nil {
-		return res.Error(err)
-	}
-	objectreferencesv1.SetObjectReference(&req.Instance.Status.RelatedObjects, *objectRef)
-
-	// Handle KubeVirt resource conditions
-	isReady := handleComponentConditions(r, req, "KubeVirt", translateKubeVirtConds(found.Status.Conditions))
-
-	upgradeDone := req.ComponentUpgradeInProgress && isReady && r.checkComponentVersion(hcoutil.KubevirtVersionEnvV, found.Status.ObservedKubeVirtVersion)
-
-	return res.SetUpgradeDone(upgradeDone)
+func (r *ReconcileHyperConverged) ensureCDI(req *common.HcoRequest) *operands.EnsureResult {
+	return operandMap["cdi"].Ensure(req)
 }
 
-func (r *ReconcileHyperConverged) ensureCDI(req *common.HcoRequest) *EnsureResult {
-	cdi := req.Instance.NewCDI()
-	res := NewEnsureResult(cdi)
-
-	key, err := client.ObjectKeyFromObject(cdi)
-	if err != nil {
-		req.Logger.Error(err, "Failed to get object key for CDI")
-	}
-
-	res.SetName(key.Name)
-	found := &cdiv1beta1.CDI{}
-	err = r.client.Get(req.Ctx, key, found)
-
-	if err != nil {
-		if apierrors.IsNotFound(err) {
-			req.Logger.Info("Creating CDI")
-			err = r.client.Create(req.Ctx, cdi)
-			if err == nil {
-				return res.SetCreated()
-			}
-		}
-		return res.Error(err)
-	}
-
-	req.Logger.Info("CDI already exists", "CDI.Namespace", found.Namespace, "CDI.Name", found.Name)
-
-	err = r.ensureKubeVirtStorageConfig(req)
-	if err != nil {
-		return res.Error(err)
-	}
-
-	err = r.ensureKubeVirtStorageRole(req)
-	if err != nil {
-		return res.Error(err)
-	}
-
-	err = r.ensureKubeVirtStorageRoleBinding(req)
-	if err != nil {
-		return res.Error(err)
-	}
-
-	existingOwners := found.GetOwnerReferences()
-
-	// Previous versions used to have HCO-operator (scope namespace)
-	// as the owner of CDI (scope cluster).
-	// It's not legal, so remove that.
-	if len(existingOwners) > 0 {
-		req.Logger.Info("CDI has owners, removing...")
-		found.SetOwnerReferences([]metav1.OwnerReference{})
-		err = r.client.Update(req.Ctx, found)
-		if err != nil {
-			req.Logger.Error(err, "Failed to remove CDI's previous owners")
-		}
-	}
-
-	if !reflect.DeepEqual(found.Spec, cdi.Spec) {
-		req.Logger.Info("Updating existing CDI' Spec to its default value")
-		cdi.Spec.DeepCopyInto(&found.Spec)
-		err = r.client.Update(req.Ctx, found)
-		if err != nil {
-			return res.Error(err)
-		}
-		return res.SetUpdated()
-	}
-
-	// Add it to the list of RelatedObjects if found
-	objectRef, err := reference.GetReference(r.scheme, found)
-	if err != nil {
-		return res.Error(err)
-	}
-	objectreferencesv1.SetObjectReference(&req.Instance.Status.RelatedObjects, *objectRef)
-
-	// Handle CDI resource conditions
-	isReady := handleComponentConditions(r, req, "CDI", found.Status.Conditions)
-
-	upgradeDone := req.ComponentUpgradeInProgress && isReady && r.checkComponentVersion(hcoutil.CdiVersionEnvV, found.Status.ObservedVersion)
-
-	return res.SetUpgradeDone(upgradeDone)
-}
-
-func (r *ReconcileHyperConverged) ensureNetworkAddons(req *common.HcoRequest) *EnsureResult {
+func (r *ReconcileHyperConverged) ensureNetworkAddons(req *common.HcoRequest) *operands.EnsureResult {
 	networkAddons := req.Instance.NewNetworkAddons()
 
-	res := NewEnsureResult(networkAddons)
+	res := operands.NewEnsureResult(networkAddons)
 	key, err := client.ObjectKeyFromObject(networkAddons)
 	if err != nil {
 		req.Logger.Error(err, "Failed to get object key for Network Addons")
@@ -1203,10 +943,10 @@ func (r *ReconcileHyperConverged) componentNotAvailable(req *common.HcoRequest, 
 	})
 }
 
-func (r *ReconcileHyperConverged) ensureKubeVirtCommonTemplateBundle(req *common.HcoRequest) *EnsureResult {
+func (r *ReconcileHyperConverged) ensureKubeVirtCommonTemplateBundle(req *common.HcoRequest) *operands.EnsureResult {
 
 	kvCTB := req.Instance.NewKubeVirtCommonTemplateBundle()
-	res := NewEnsureResult(kvCTB)
+	res := operands.NewEnsureResult(kvCTB)
 	if !r.clusterInfo.IsOpenshift() { // SSP operators Only supported in OpenShift. Ignore in K8s.
 		return res.SetUpgradeDone(true)
 	}
@@ -1317,9 +1057,9 @@ func newKubeVirtNodeLabellerBundleForCR(cr *hcov1beta1.HyperConverged, namespace
 	}
 }
 
-func (r *ReconcileHyperConverged) ensureKubeVirtNodeLabellerBundle(req *common.HcoRequest) *EnsureResult {
+func (r *ReconcileHyperConverged) ensureKubeVirtNodeLabellerBundle(req *common.HcoRequest) *operands.EnsureResult {
 	kvNLB := newKubeVirtNodeLabellerBundleForCR(req.Instance, req.Namespace)
-	res := NewEnsureResult(kvNLB)
+	res := operands.NewEnsureResult(kvNLB)
 	if !r.clusterInfo.IsOpenshift() { // SSP operators Only supported in OpenShift. Ignore in K8s.
 		return res.SetUpgradeDone(true)
 	}
@@ -1400,9 +1140,9 @@ func newIMSConfigForCR(cr *hcov1beta1.HyperConverged, namespace string) *corev1.
 	}
 }
 
-func (r *ReconcileHyperConverged) ensureIMSConfig(req *common.HcoRequest) *EnsureResult {
+func (r *ReconcileHyperConverged) ensureIMSConfig(req *common.HcoRequest) *operands.EnsureResult {
 	imsConfig := newIMSConfigForCR(req.Instance, req.Namespace)
-	res := NewEnsureResult(imsConfig)
+	res := operands.NewEnsureResult(imsConfig)
 	if os.Getenv("CONVERSION_CONTAINER") == "" {
 		return res.Error(errors.New("ims-conversion-container not specified"))
 	}
@@ -1468,9 +1208,9 @@ func (r *ReconcileHyperConverged) ensureIMSConfig(req *common.HcoRequest) *Ensur
 	return res.SetUpgradeDone(req.ComponentUpgradeInProgress)
 }
 
-func (r *ReconcileHyperConverged) ensureVMImport(req *common.HcoRequest) *EnsureResult {
+func (r *ReconcileHyperConverged) ensureVMImport(req *common.HcoRequest) *operands.EnsureResult {
 	vmImport := newVMImportForCR(req.Instance)
-	res := NewEnsureResult(vmImport)
+	res := operands.NewEnsureResult(vmImport)
 
 	key, err := client.ObjectKeyFromObject(vmImport)
 	if err != nil {
@@ -1617,9 +1357,9 @@ func newKubeVirtTemplateValidatorForCR(cr *hcov1beta1.HyperConverged, namespace 
 	}
 }
 
-func (r *ReconcileHyperConverged) ensureKubeVirtTemplateValidator(req *common.HcoRequest) *EnsureResult {
+func (r *ReconcileHyperConverged) ensureKubeVirtTemplateValidator(req *common.HcoRequest) *operands.EnsureResult {
 	kvTV := newKubeVirtTemplateValidatorForCR(req.Instance, req.Namespace)
-	res := NewEnsureResult(kvTV)
+	res := operands.NewEnsureResult(kvTV)
 	if !r.clusterInfo.IsOpenshift() { // SSP operators Only supported in OpenShift. Ignore in K8s.
 		return res.SetUpgradeDone(true)
 	}
@@ -1680,175 +1420,6 @@ func (r *ReconcileHyperConverged) ensureKubeVirtTemplateValidator(req *common.Hc
 	return res.SetUpgradeDone(req.ComponentUpgradeInProgress && upgradeInProgress)
 }
 
-func newKubeVirtStorageRoleForCR(cr *hcov1beta1.HyperConverged, namespace string) *rbacv1.Role {
-	labels := map[string]string{
-		"app": cr.Name,
-	}
-	return &rbacv1.Role{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "hco.kubevirt.io:config-reader",
-			Labels:    labels,
-			Namespace: namespace,
-		},
-		Rules: []rbacv1.PolicyRule{
-			{
-				APIGroups:     []string{""},
-				Resources:     []string{"configmaps"},
-				ResourceNames: []string{"kubevirt-storage-class-defaults"},
-				Verbs:         []string{"get", "watch", "list"},
-			},
-		},
-	}
-}
-
-func newKubeVirtStorageRoleBindingForCR(cr *hcov1beta1.HyperConverged, namespace string) *rbacv1.RoleBinding {
-	labels := map[string]string{
-		"app": cr.Name,
-	}
-	return &rbacv1.RoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "hco.kubevirt.io:config-reader",
-			Labels:    labels,
-			Namespace: namespace,
-		},
-		RoleRef: rbacv1.RoleRef{
-			APIGroup: "rbac.authorization.k8s.io",
-			Kind:     "Role",
-			Name:     "hco.kubevirt.io:config-reader",
-		},
-		Subjects: []rbacv1.Subject{
-			{
-				APIGroup: "rbac.authorization.k8s.io",
-				Kind:     "Group",
-				Name:     "system:authenticated",
-			},
-		},
-	}
-}
-
-func newKubeVirtStorageConfigForCR(cr *hcov1beta1.HyperConverged, namespace string) *corev1.ConfigMap {
-	localSC := "local-sc"
-	if *(&cr.Spec.LocalStorageClassName) != "" {
-		localSC = *(&cr.Spec.LocalStorageClassName)
-	}
-
-	labels := map[string]string{
-		hcoutil.AppLabel: cr.Name,
-	}
-	return &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kubevirt-storage-class-defaults",
-			Labels:    labels,
-			Namespace: namespace,
-		},
-		Data: map[string]string{
-			"accessMode":            "ReadWriteOnce",
-			"volumeMode":            "Filesystem",
-			localSC + ".accessMode": "ReadWriteOnce",
-			localSC + ".volumeMode": "Filesystem",
-		},
-	}
-}
-
-func (r *ReconcileHyperConverged) ensureKubeVirtStorageRole(req *common.HcoRequest) error {
-	kubevirtStorageRole := newKubeVirtStorageRoleForCR(req.Instance, req.Namespace)
-	if err := controllerutil.SetControllerReference(req.Instance, kubevirtStorageRole, r.scheme); err != nil {
-		return err
-	}
-
-	key, err := client.ObjectKeyFromObject(kubevirtStorageRole)
-	if err != nil {
-		req.Logger.Error(err, "Failed to get object key for kubevirt storage role")
-	}
-
-	found := &rbacv1.Role{}
-	err = r.client.Get(req.Ctx, key, found)
-	if err != nil && apierrors.IsNotFound(err) {
-		req.Logger.Info("Creating kubevirt storage role")
-		return r.client.Create(req.Ctx, kubevirtStorageRole)
-	}
-
-	if err != nil {
-		return err
-	}
-
-	req.Logger.Info("KubeVirt storage role already exists", "KubeVirtConfig.Namespace", found.Namespace, "KubeVirtConfig.Name", found.Name)
-	// Add it to the list of RelatedObjects if found
-	objectRef, err := reference.GetReference(r.scheme, found)
-	if err != nil {
-		return err
-	}
-	objectreferencesv1.SetObjectReference(&req.Instance.Status.RelatedObjects, *objectRef)
-
-	return nil
-}
-
-func (r *ReconcileHyperConverged) ensureKubeVirtStorageRoleBinding(req *common.HcoRequest) error {
-	kubevirtStorageRoleBinding := newKubeVirtStorageRoleBindingForCR(req.Instance, req.Namespace)
-	if err := controllerutil.SetControllerReference(req.Instance, kubevirtStorageRoleBinding, r.scheme); err != nil {
-		return err
-	}
-
-	key, err := client.ObjectKeyFromObject(kubevirtStorageRoleBinding)
-	if err != nil {
-		req.Logger.Error(err, "Failed to get object key for kubevirt storage rolebinding")
-	}
-
-	found := &rbacv1.RoleBinding{}
-	err = r.client.Get(req.Ctx, key, found)
-	if err != nil && apierrors.IsNotFound(err) {
-		req.Logger.Info("Creating kubevirt storage rolebinding")
-		return r.client.Create(req.Ctx, kubevirtStorageRoleBinding)
-	}
-
-	if err != nil {
-		return err
-	}
-
-	req.Logger.Info("KubeVirt storage rolebinding already exists", "KubeVirtConfig.Namespace", found.Namespace, "KubeVirtConfig.Name", found.Name)
-	// Add it to the list of RelatedObjects if found
-	objectRef, err := reference.GetReference(r.scheme, found)
-	if err != nil {
-		return err
-	}
-	objectreferencesv1.SetObjectReference(&req.Instance.Status.RelatedObjects, *objectRef)
-
-	return nil
-}
-
-func (r *ReconcileHyperConverged) ensureKubeVirtStorageConfig(req *common.HcoRequest) error {
-	kubevirtStorageConfig := newKubeVirtStorageConfigForCR(req.Instance, req.Namespace)
-	if err := controllerutil.SetControllerReference(req.Instance, kubevirtStorageConfig, r.scheme); err != nil {
-		return err
-	}
-
-	key, err := client.ObjectKeyFromObject(kubevirtStorageConfig)
-	if err != nil {
-		req.Logger.Error(err, "Failed to get object key for kubevirt storage config")
-	}
-
-	found := &corev1.ConfigMap{}
-	err = r.client.Get(req.Ctx, key, found)
-	if err != nil && apierrors.IsNotFound(err) {
-		req.Logger.Info("Creating kubevirt storage config")
-		return r.client.Create(req.Ctx, kubevirtStorageConfig)
-	}
-
-	if err != nil {
-		return err
-	}
-
-	req.Logger.Info("KubeVirt storage config already exists", "KubeVirtConfig.Namespace", found.Namespace, "KubeVirtConfig.Name", found.Name)
-	// Add it to the list of RelatedObjects if found
-	objectRef, err := reference.GetReference(r.scheme, found)
-	if err != nil {
-		return err
-	}
-	objectreferencesv1.SetObjectReference(&req.Instance.Status.RelatedObjects, *objectRef)
-
-	return nil
-}
-
 func newKubeVirtMetricsAggregationForCR(cr *hcov1beta1.HyperConverged, namespace string) *sspv1.KubevirtMetricsAggregation {
 	labels := map[string]string{
 		hcoutil.AppLabel: cr.Name,
@@ -1862,9 +1433,9 @@ func newKubeVirtMetricsAggregationForCR(cr *hcov1beta1.HyperConverged, namespace
 	}
 }
 
-func (r *ReconcileHyperConverged) ensureKubeVirtMetricsAggregation(req *common.HcoRequest) *EnsureResult {
+func (r *ReconcileHyperConverged) ensureKubeVirtMetricsAggregation(req *common.HcoRequest) *operands.EnsureResult {
 	kubevirtMetricsAggregation := newKubeVirtMetricsAggregationForCR(req.Instance, req.Namespace)
-	res := NewEnsureResult(kubevirtMetricsAggregation)
+	res := operands.NewEnsureResult(kubevirtMetricsAggregation)
 	if !r.clusterInfo.IsOpenshift() { // SSP operators Only supported in OpenShift. Ignore in K8s.
 		return res.SetUpgradeDone(true)
 	}
