+++ pkg/controller/operands/cdi_test.go
@@ -1,0 +1,350 @@
+package operands
+
+import (
+	"context"
+	"fmt"
+	hcov1beta1 "github.com/kubevirt/hyperconverged-cluster-operator/pkg/apis/hco/v1beta1"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/common"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/commonTestUtils"
+	hcoutil "github.com/kubevirt/hyperconverged-cluster-operator/pkg/util"
+	conditionsv1 "github.com/openshift/custom-resource-status/conditions/v1"
+	"github.com/openshift/custom-resource-status/testlib"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/tools/reference"
+	cdiv1beta1 "kubevirt.io/containerized-data-importer/pkg/apis/core/v1beta1"
+	"k8s.io/apimachinery/pkg/types"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+var _ = Describe("CDI Operand", func() {
+	Context("CDI", func() {
+		var (
+			hco *hcov1beta1.HyperConverged
+			req *common.HcoRequest
+		)
+
+		BeforeEach(func() {
+			hco = commonTestUtils.NewHco()
+			req = commonTestUtils.NewReq(hco)
+		})
+
+		It("should create if not present", func() {
+			expectedResource := hco.NewCDI()
+			cl := commonTestUtils.InitClient([]runtime.Object{})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			res := handler.Ensure(req)
+			Expect(res.UpgradeDone).To(BeFalse())
+			Expect(res.Err).To(BeNil())
+
+			foundResource := &cdiv1beta1.CDI{}
+			Expect(
+				cl.Get(context.TODO(),
+					types.NamespacedName{Name: expectedResource.Name, Namespace: expectedResource.Namespace},
+					foundResource),
+			).To(BeNil())
+			Expect(foundResource.Name).To(Equal(expectedResource.Name))
+			Expect(foundResource.Labels).Should(HaveKeyWithValue(hcoutil.AppLabel, commonTestUtils.Name))
+			Expect(foundResource.Namespace).To(Equal(expectedResource.Namespace))
+		})
+
+		It("should find if present", func() {
+			expectedResource := hco.NewCDI()
+			expectedResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/namespaces/%s/dummies/%s", expectedResource.Namespace, expectedResource.Name)
+			cl := commonTestUtils.InitClient([]runtime.Object{hco, expectedResource})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			res := handler.Ensure(req)
+			Expect(res.UpgradeDone).To(BeFalse())
+			Expect(res.Err).To(BeNil())
+
+			// Check HCO's status
+			Expect(hco.Status.RelatedObjects).To(Not(BeNil()))
+			objectRef, err := reference.GetReference(handler.Scheme, expectedResource)
+			Expect(err).To(BeNil())
+			// ObjectReference should have been added
+			Expect(hco.Status.RelatedObjects).To(ContainElement(*objectRef))
+			// Check conditions
+			Expect(req.Conditions[conditionsv1.ConditionAvailable]).To(testlib.RepresentCondition(conditionsv1.Condition{
+				Type:    conditionsv1.ConditionAvailable,
+				Status:  corev1.ConditionFalse,
+				Reason:  "CDIConditions",
+				Message: "CDI resource has no conditions",
+			}))
+			Expect(req.Conditions[conditionsv1.ConditionProgressing]).To(testlib.RepresentCondition(conditionsv1.Condition{
+				Type:    conditionsv1.ConditionProgressing,
+				Status:  corev1.ConditionTrue,
+				Reason:  "CDIConditions",
+				Message: "CDI resource has no conditions",
+			}))
+			Expect(req.Conditions[conditionsv1.ConditionUpgradeable]).To(testlib.RepresentCondition(conditionsv1.Condition{
+				Type:    conditionsv1.ConditionUpgradeable,
+				Status:  corev1.ConditionFalse,
+				Reason:  "CDIConditions",
+				Message: "CDI resource has no conditions",
+			}))
+		})
+
+		It("should set default UninstallStrategy if missing", func() {
+			expectedResource := hco.NewCDI()
+			expectedResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/%s/dummies/%s", expectedResource.Namespace, expectedResource.Name)
+			missingUSResource := hco.NewCDI()
+			missingUSResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/%s/dummies/%s", missingUSResource.Namespace, missingUSResource.Name)
+			missingUSResource.Spec.UninstallStrategy = nil
+
+			cl := commonTestUtils.InitClient([]runtime.Object{hco, missingUSResource})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			res := handler.Ensure(req)
+			Expect(res.UpgradeDone).To(BeFalse())
+			Expect(res.Updated).To(BeTrue())
+			Expect(res.Err).To(BeNil())
+
+			foundResource := &cdiv1beta1.CDI{}
+			Expect(
+				cl.Get(context.TODO(),
+					types.NamespacedName{Name: expectedResource.Name, Namespace: expectedResource.Namespace},
+					foundResource),
+			).To(BeNil())
+			Expect(*foundResource.Spec.UninstallStrategy).To(Equal(*expectedResource.Spec.UninstallStrategy))
+		})
+
+		It("should add node placement if missing in CDI", func() {
+			existingResource := hco.NewCDI()
+
+			hco.Spec.Infra = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
+			hco.Spec.Workloads = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
+
+			cl := commonTestUtils.InitClient([]runtime.Object{hco, existingResource})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			res := handler.Ensure(req)
+			Expect(res.UpgradeDone).To(BeFalse())
+			Expect(res.Updated).To(BeTrue())
+			Expect(res.Err).To(BeNil())
+
+			foundResource := &cdiv1beta1.CDI{}
+			Expect(
+				cl.Get(context.TODO(),
+					types.NamespacedName{Name: existingResource.Name, Namespace: existingResource.Namespace},
+					foundResource),
+			).To(BeNil())
+
+			Expect(existingResource.Spec.Infra.Affinity).To(BeNil())
+			Expect(existingResource.Spec.Infra.Tolerations).To(BeEmpty())
+			Expect(existingResource.Spec.Infra.NodeSelector).To(BeNil())
+			Expect(existingResource.Spec.Workloads.Affinity).To(BeNil())
+			Expect(existingResource.Spec.Workloads.Tolerations).To(BeEmpty())
+			Expect(existingResource.Spec.Workloads.NodeSelector).To(BeNil())
+
+			Expect(foundResource.Spec.Infra.Affinity).ToNot(BeNil())
+			Expect(foundResource.Spec.Infra.NodeSelector["key1"]).Should(Equal("value1"))
+			Expect(foundResource.Spec.Infra.NodeSelector["key2"]).Should(Equal("value2"))
+
+			Expect(foundResource.Spec.Workloads).ToNot(BeNil())
+			Expect(foundResource.Spec.Workloads.Tolerations).Should(Equal(hco.Spec.Workloads.NodePlacement.Tolerations))
+
+			Expect(req.Conditions).To(BeEmpty())
+		})
+
+		It("should remove node placement if missing in HCO CR", func() {
+
+			hcoNodePlacement := commonTestUtils.NewHco()
+			hcoNodePlacement.Spec.Infra = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
+			hcoNodePlacement.Spec.Workloads = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
+			existingResource := hcoNodePlacement.NewCDI()
+
+			cl := commonTestUtils.InitClient([]runtime.Object{hco, existingResource})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			res := handler.Ensure(req)
+			Expect(res.UpgradeDone).To(BeFalse())
+			Expect(res.Updated).To(BeTrue())
+			Expect(res.Err).To(BeNil())
+
+			foundResource := &cdiv1beta1.CDI{}
+			Expect(
+				cl.Get(context.TODO(),
+					types.NamespacedName{Name: existingResource.Name, Namespace: existingResource.Namespace},
+					foundResource),
+			).To(BeNil())
+
+			Expect(existingResource.Spec.Infra.Affinity).ToNot(BeNil())
+			Expect(existingResource.Spec.Infra.Tolerations).ToNot(BeEmpty())
+			Expect(existingResource.Spec.Infra.NodeSelector).ToNot(BeNil())
+			Expect(existingResource.Spec.Workloads.Affinity).ToNot(BeNil())
+			Expect(existingResource.Spec.Workloads.Tolerations).ToNot(BeEmpty())
+			Expect(existingResource.Spec.Workloads.NodeSelector).ToNot(BeNil())
+
+			Expect(foundResource.Spec.Infra.Affinity).To(BeNil())
+			Expect(foundResource.Spec.Infra.Tolerations).To(BeEmpty())
+			Expect(foundResource.Spec.Infra.NodeSelector).To(BeNil())
+			Expect(foundResource.Spec.Workloads.Affinity).To(BeNil())
+			Expect(foundResource.Spec.Workloads.Tolerations).To(BeEmpty())
+			Expect(foundResource.Spec.Workloads.NodeSelector).To(BeNil())
+
+			Expect(req.Conditions).To(BeEmpty())
+		})
+
+		It("should modify node placement according to HCO CR", func() {
+			hco.Spec.Infra = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
+			hco.Spec.Workloads = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
+			existingResource := hco.NewCDI()
+
+			// now, modify HCO's node placement
+			seconds3 := int64(3)
+			hco.Spec.Infra.NodePlacement.Tolerations = append(hco.Spec.Infra.NodePlacement.Tolerations, corev1.Toleration{
+				Key: "key3", Operator: "operator3", Value: "value3", Effect: "effect3", TolerationSeconds: &seconds3,
+			})
+
+			hco.Spec.Workloads.NodePlacement.NodeSelector["key1"] = "something else"
+
+			cl := commonTestUtils.InitClient([]runtime.Object{hco, existingResource})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			res := handler.Ensure(req)
+			Expect(res.UpgradeDone).To(BeFalse())
+			Expect(res.Updated).To(BeTrue())
+			Expect(res.Err).To(BeNil())
+
+			foundResource := &cdiv1beta1.CDI{}
+			Expect(
+				cl.Get(context.TODO(),
+					types.NamespacedName{Name: existingResource.Name, Namespace: existingResource.Namespace},
+					foundResource),
+			).To(BeNil())
+
+			Expect(existingResource.Spec.Infra.Tolerations).To(HaveLen(2))
+			Expect(existingResource.Spec.Workloads.NodeSelector["key1"]).Should(Equal("value1"))
+
+			Expect(foundResource.Spec.Infra.Tolerations).To(HaveLen(3))
+			Expect(foundResource.Spec.Workloads.NodeSelector["key1"]).Should(Equal("something else"))
+
+			Expect(req.Conditions).To(BeEmpty())
+		})
+
+		It("should handle conditions", func() {
+			expectedResource := hco.NewCDI()
+			expectedResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/namespaces/%s/dummies/%s", expectedResource.Namespace, expectedResource.Name)
+			expectedResource.Status.Conditions = []conditionsv1.Condition{
+				conditionsv1.Condition{
+					Type:    conditionsv1.ConditionAvailable,
+					Status:  corev1.ConditionFalse,
+					Reason:  "Foo",
+					Message: "Bar",
+				},
+				conditionsv1.Condition{
+					Type:    conditionsv1.ConditionProgressing,
+					Status:  corev1.ConditionTrue,
+					Reason:  "Foo",
+					Message: "Bar",
+				},
+				conditionsv1.Condition{
+					Type:    conditionsv1.ConditionDegraded,
+					Status:  corev1.ConditionTrue,
+					Reason:  "Foo",
+					Message: "Bar",
+				},
+			}
+			cl := commonTestUtils.InitClient([]runtime.Object{hco, expectedResource})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			res := handler.Ensure(req)
+			Expect(res.UpgradeDone).To(BeFalse())
+			Expect(res.Err).To(BeNil())
+
+			// Check HCO's status
+			Expect(hco.Status.RelatedObjects).To(Not(BeNil()))
+			objectRef, err := reference.GetReference(handler.Scheme, expectedResource)
+			Expect(err).To(BeNil())
+			// ObjectReference should have been added
+			Expect(hco.Status.RelatedObjects).To(ContainElement(*objectRef))
+			// Check conditions
+			Expect(req.Conditions[conditionsv1.ConditionAvailable]).To(testlib.RepresentCondition(conditionsv1.Condition{
+				Type:    conditionsv1.ConditionAvailable,
+				Status:  corev1.ConditionFalse,
+				Reason:  "CDINotAvailable",
+				Message: "CDI is not available: Bar",
+			}))
+			Expect(req.Conditions[conditionsv1.ConditionProgressing]).To(testlib.RepresentCondition(conditionsv1.Condition{
+				Type:    conditionsv1.ConditionProgressing,
+				Status:  corev1.ConditionTrue,
+				Reason:  "CDIProgressing",
+				Message: "CDI is progressing: Bar",
+			}))
+			Expect(req.Conditions[conditionsv1.ConditionUpgradeable]).To(testlib.RepresentCondition(conditionsv1.Condition{
+				Type:    conditionsv1.ConditionUpgradeable,
+				Status:  corev1.ConditionFalse,
+				Reason:  "CDIProgressing",
+				Message: "CDI is progressing: Bar",
+			}))
+			Expect(req.Conditions[conditionsv1.ConditionDegraded]).To(testlib.RepresentCondition(conditionsv1.Condition{
+				Type:    conditionsv1.ConditionDegraded,
+				Status:  corev1.ConditionTrue,
+				Reason:  "CDIDegraded",
+				Message: "CDI is degraded: Bar",
+			}))
+		})
+	})
+
+	Context("KubeVirt Storage Config", func() {
+		var hco *hcov1beta1.HyperConverged
+		var req *common.HcoRequest
+
+		BeforeEach(func() {
+			hco = commonTestUtils.NewHco()
+			req = commonTestUtils.NewReq(hco)
+		})
+
+		It("should create if not present", func() {
+			expectedResource := NewKubeVirtStorageConfigForCR(hco, commonTestUtils.Namespace)
+			cl := commonTestUtils.InitClient([]runtime.Object{})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			err := handler.ensureKubeVirtStorageConfig(req)
+			Expect(err).To(BeNil())
+
+			foundResource := &corev1.ConfigMap{}
+			Expect(
+				cl.Get(context.TODO(),
+					types.NamespacedName{Name: expectedResource.Name, Namespace: expectedResource.Namespace},
+					foundResource),
+			).To(BeNil())
+			Expect(foundResource.Name).To(Equal(expectedResource.Name))
+			Expect(foundResource.Labels).Should(HaveKeyWithValue(hcoutil.AppLabel, commonTestUtils.Name))
+			Expect(foundResource.Namespace).To(Equal(expectedResource.Namespace))
+		})
+
+		It("should find if present", func() {
+			expectedResource := NewKubeVirtStorageConfigForCR(hco, commonTestUtils.Namespace)
+			expectedResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/namespaces/%s/dummies/%s", expectedResource.Namespace, expectedResource.Name)
+			cl := commonTestUtils.InitClient([]runtime.Object{hco, expectedResource})
+			handler := &CdiHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
+			err := handler.ensureKubeVirtStorageConfig(req)
+			Expect(err).To(BeNil())
+
+			// Check HCO's status
+			Expect(hco.Status.RelatedObjects).To(Not(BeNil()))
+			objectRef, err := reference.GetReference(handler.Scheme, expectedResource)
+			Expect(err).To(BeNil())
+			// ObjectReference should have been added
+			Expect(hco.Status.RelatedObjects).To(ContainElement(*objectRef))
+		})
+
+		It("volumeMode should be filesystem when platform is baremetal", func() {
+			hco.Spec.BareMetalPlatform = true
+
+			expectedResource := NewKubeVirtStorageConfigForCR(hco, commonTestUtils.Namespace)
+			Expect(expectedResource.Data["volumeMode"]).To(Equal("Filesystem"))
+		})
+
+		It("volumeMode should be filesystem when platform is not baremetal", func() {
+			hco.Spec.BareMetalPlatform = false
+
+			expectedResource := NewKubeVirtStorageConfigForCR(hco, commonTestUtils.Namespace)
+			Expect(expectedResource.Data["volumeMode"]).To(Equal("Filesystem"))
+		})
+
+		It("local storage class name should be available when specified", func() {
+			hco.Spec.LocalStorageClassName = "local"
+
+			expectedResource := NewKubeVirtStorageConfigForCR(hco, commonTestUtils.Namespace)
+			Expect(expectedResource.Data["local.accessMode"]).To(Equal("ReadWriteOnce"))
+			Expect(expectedResource.Data["local.volumeMode"]).To(Equal("Filesystem"))
+		})
+	})
+})
