+++ pkg/controller/operands/ensure_result_test.go
@@ -1,4 +1,4 @@
-package hyperconverged
+package operands
 
 import (
 	"errors"
