+++ pkg/controller/operands/operand.go
@@ -1,0 +1,118 @@
+package operands
+
+import (
+	"fmt"
+	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/common"
+	conditionsv1 "github.com/openshift/custom-resource-status/conditions/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"os"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+type Operand interface {
+	Ensure(req *common.HcoRequest) *EnsureResult
+}
+
+type genericOperand struct {
+	Client client.Client
+	Scheme *runtime.Scheme
+}
+
+// handleComponentConditions - read and process a sub-component conditions.
+// returns true if the the conditions indicates "ready" state and false if not.
+func handleComponentConditions(req *common.HcoRequest, component string, componentConds []conditionsv1.Condition) (isReady bool) {
+	isReady = true
+	if len(componentConds) == 0 {
+		isReady = false
+		reason := fmt.Sprintf("%sConditions", component)
+		message := fmt.Sprintf("%s resource has no conditions", component)
+		req.Logger.Info(fmt.Sprintf("%s's resource is not reporting Conditions on it's Status", component))
+		req.Conditions.SetStatusCondition(conditionsv1.Condition{
+			Type:    conditionsv1.ConditionAvailable,
+			Status:  corev1.ConditionFalse,
+			Reason:  reason,
+			Message: message,
+		})
+		req.Conditions.SetStatusCondition(conditionsv1.Condition{
+			Type:    conditionsv1.ConditionProgressing,
+			Status:  corev1.ConditionTrue,
+			Reason:  reason,
+			Message: message,
+		})
+		req.Conditions.SetStatusCondition(conditionsv1.Condition{
+			Type:    conditionsv1.ConditionUpgradeable,
+			Status:  corev1.ConditionFalse,
+			Reason:  reason,
+			Message: message,
+		})
+	} else {
+		foundAvailableCond := false
+		foundProgressingCond := false
+		foundDegradedCond := false
+		for _, condition := range componentConds {
+			switch condition.Type {
+			case conditionsv1.ConditionAvailable:
+				foundAvailableCond = true
+				if condition.Status == corev1.ConditionFalse {
+					isReady = false
+					msg := fmt.Sprintf("%s is not available: %v", component, string(condition.Message))
+					componentNotAvailable(req, component, msg)
+				}
+			case conditionsv1.ConditionProgressing:
+				foundProgressingCond = true
+				if condition.Status == corev1.ConditionTrue {
+					isReady = false
+					req.Logger.Info(fmt.Sprintf("%s is 'Progressing'", component))
+					req.Conditions.SetStatusCondition(conditionsv1.Condition{
+						Type:    conditionsv1.ConditionProgressing,
+						Status:  corev1.ConditionTrue,
+						Reason:  fmt.Sprintf("%sProgressing", component),
+						Message: fmt.Sprintf("%s is progressing: %v", component, string(condition.Message)),
+					})
+					req.Conditions.SetStatusCondition(conditionsv1.Condition{
+						Type:    conditionsv1.ConditionUpgradeable,
+						Status:  corev1.ConditionFalse,
+						Reason:  fmt.Sprintf("%sProgressing", component),
+						Message: fmt.Sprintf("%s is progressing: %v", component, string(condition.Message)),
+					})
+				}
+			case conditionsv1.ConditionDegraded:
+				foundDegradedCond = true
+				if condition.Status == corev1.ConditionTrue {
+					isReady = false
+					req.Logger.Info(fmt.Sprintf("%s is 'Degraded'", component))
+					req.Conditions.SetStatusCondition(conditionsv1.Condition{
+						Type:    conditionsv1.ConditionDegraded,
+						Status:  corev1.ConditionTrue,
+						Reason:  fmt.Sprintf("%sDegraded", component),
+						Message: fmt.Sprintf("%s is degraded: %v", component, string(condition.Message)),
+					})
+				}
+			}
+		}
+
+		if !foundAvailableCond {
+			componentNotAvailable(req, component, `missing "Available" condition`)
+		}
+
+		isReady = isReady && foundAvailableCond && foundProgressingCond && foundDegradedCond
+	}
+
+	return isReady
+}
+
+func componentNotAvailable(req *common.HcoRequest, component string, msg string) {
+	req.Logger.Info(fmt.Sprintf("%s is not 'Available'", component))
+	req.Conditions.SetStatusCondition(conditionsv1.Condition{
+		Type:    conditionsv1.ConditionAvailable,
+		Status:  corev1.ConditionFalse,
+		Reason:  fmt.Sprintf("%sNotAvailable", component),
+		Message: msg,
+	})
+}
+
+func checkComponentVersion(versionEnvName, actualVersion string) bool {
+	expectedVersion := os.Getenv(versionEnvName)
+	return expectedVersion != "" && expectedVersion == actualVersion
+}
