+++ pkg/controller/operands/operands_suite_test.go
@@ -1,0 +1,13 @@
+package operands_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestHyperconverged(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Operands Suite")
+}
