+++ pkg/controller/operands/testUtils_test.go
@@ -1,0 +1,47 @@
+package operands_test
+
+import (
+    "context"
+    hcov1beta1 "github.com/kubevirt/hyperconverged-cluster-operator/pkg/apis/hco/v1beta1"
+    "github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/common"
+    metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+    "sigs.k8s.io/controller-runtime/pkg/reconcile"
+    "k8s.io/apimachinery/pkg/types"
+    logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
+)
+
+// name and namespace of our primary resource
+const (
+    name      = "kubevirt-hyperconverged"
+    namespace = "kubevirt-hyperconverged"
+)
+
+var (
+    request = reconcile.Request{
+        NamespacedName: types.NamespacedName{
+            Name:      name,
+            Namespace: namespace,
+        },
+    }
+    log = logf.Log.WithName("controller_hyperconverged")
+)
+
+func newHco() *hcov1beta1.HyperConverged {
+    return &hcov1beta1.HyperConverged{
+        ObjectMeta: metav1.ObjectMeta{
+            Name:      name,
+            Namespace: namespace,
+        },
+        Spec: hcov1beta1.HyperConvergedSpec{},
+    }
+}
+
+func newReq(inst *hcov1beta1.HyperConverged) *common.HcoRequest {
+    return &common.HcoRequest{
+        Request:    request,
+        Logger:     log,
+        Conditions: common.NewHcoConditions(),
+        Ctx:        context.TODO(),
+        Instance:   inst,
+    }
+}
